#30 se modifica la vista para poder ver en que pelicula trabajo y que rol interpreto. se agrega un boton que permite borrar el personaje en el que participa.

+++ IMDB/IMDB/Controllers/MovieController.cs
@@ -79,7 +79,7 @@ namespace IMDB.Controllers
             return View(movieViewModel);// envio la lista de tipo MovieViewModel que contiene las peliculas con ese modelo 
         }
 
-        public ActionResult Details(long Id)
+        public ActionResult Details(int Id)
         {
             // traer pelicula >> metodo para buscar por id
             var movieById = db.GetMovieById(Id);
